Rename profile-service to profile-service-sdk
