+++ apps/rwa/package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "@auth0/nextjs-auth0": "^3.2.0",
     "next": "^13.4.19",
-    "profile-service": "workspace:*",
+    "profile-service-sdk": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "ui-components": "workspace:*"
