+++ apps/spa/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@auth0/auth0-react": "^2.2.3",
-    "profile-service": "workspace:*",
+    "profile-service-sdk": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.17.0",
