+++ apps/spa/src/routes/profile/index.tsx
@@ -1,5 +1,8 @@
 import { useAuth0, withAuthenticationRequired } from "@auth0/auth0-react";
-import { Profile, getProfile } from "profile-service";
+import {
+  Profile,
+  getProfile,
+} from "../../../../../packages/profile-service-sdk/src";
 import { useEffect, useState } from "react";
 
 function ProfileRoute(): JSX.Element {
